+++ CHANGELOG.md
@@ -4,6 +4,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
 
+- Fix #2276: Improve keyboard shortcuts (h, f) in 3D model viewer
 - Fix #1955: Edit user without being forced to enter a password + add a password regex
 
 ## v4.4.8 - 2025-04-08 - 01de0477b -